A failing test: "\r\n" straddling the buffer end.
If the buffer obtained from the inner reader ends with CR and
then the next portion starts with LF and the boundary, the check
for the last two bytes being "\r\n" will fail, since the CR is
no longer visible - it has (erroneously) been sent to the caller as
part of the form part's contents. Thus, the LF will also (again,
erroneously) be sent to the caller.